requirements.txt
flask==2.0.1
python-dotenv==0.19.0
okx-api==0.1.0
pandas==1.3.3
numpy==1.21.2
requests==2.26.0
python-binance==1.0.15
ccxt==1.91.9 